Fix: remove hardcoded database path
Move the neo4j derivation into a function so it can have the db path
as an argument. This can then be overridden by another flake that
depends on neo4j.

Use "/homeless-shelter" by default to force specify the db path. This
the system to have multiple dbs (one for each project) despite neo4j-community only allowing
a single one.