Clean: move neo4j configuration to a wrapper function
This allows for changing conf values without having to reinstall
neo4j. Instead a link farm is created then the original neo4j
install's conf is copied to this wrapper and the copied conf is
modified.

This also allows for adding plugins which will be described in other
nix files to the wrapper's plugin directory. Multiple projects can
then depend on separate plugins without environment leaking.