Fix: handle plugins that don't need unrestricted access
As the comment said: if you have a mix of plugins, some of which need
unrestricted access and others that don't, then you end up with a list
of package names and empty strings. When this list is reduced with the
string concatenation, you'd have multiple separators where ever the
empty strings were.